requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
web3==6.11.3
openpyxl==3.1.2

# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# PostgreSQL
psycopg2-binary>=2.9.9
